chore: reorganize solution directory
