+++ jobs/Backend/Task/src/ExchangeRateUpdater/Contracts/ExchangeRate.cs
@@ -1,4 +1,4 @@
-﻿namespace ExchangeRateUpdater
+﻿namespace ExchangeRateUpdater.Contracts
 {
     public class ExchangeRate
     {
